refactor: organise dir
