+++ client/src/features/viewers/ViewWrapper.jsx
@@ -2,7 +2,7 @@
 import { useEffect, useMemo, useState } from 'react';
 
 import { useSocket } from '../../common/context/socketContext';
-import useSubscription from '../../common/context/useSubscription';
+import useSubscription from '../../common/hooks/useSubscription';
 import useEvent from '../../common/hooks-query/useEvent';
 import useEventsList from '../../common/hooks-query/useEventsList';
 import useViewSettings from '../../common/hooks-query/useViewSettings';
